#9 [FIX] 다른 유저의 리뷰는 delete/patch 못하게 핸들링

+++ src/controllers/review.ts
@@ -1,7 +1,8 @@
 import { RequestHandler } from 'express';
 import { Review, UserReviewLike } from '../models';
+import { isUserOwnerOfReview } from '../services/review';
 
-const user_id = 1; //더미 데이터
+const user_id = '1'; //더미 데이터
 
 const getReviews: RequestHandler = async (req, res, next) => {
     try {
@@ -46,6 +47,11 @@ const postReviews: RequestHandler = async (req, res, next) => {
 
 const deleteReviews: RequestHandler = async (req, res, next) => {
     try {
+        if (!(await isUserOwnerOfReview(user_id, req.params.reviewId))) {
+            const error = new Error('다른 유저의 리뷰는 삭제할 수 없습니다.');
+            error.status = 400;
+            next(error);
+        }
         await Review.destroy({
             where: {
                 review_id: req.params.reviewId,
@@ -59,6 +65,11 @@ const deleteReviews: RequestHandler = async (req, res, next) => {
 
 const updateReviews: RequestHandler = async (req, res, next) => {
     try {
+        if (!(await isUserOwnerOfReview(user_id, req.params.reviewId))) {
+            const error = new Error('다른 유저의 리뷰는 수정할 수 없습니다.');
+            error.status = 400;
+            next(error);
+        }
         await Review.update(
             {
                 content: req.body.content,
@@ -77,10 +88,7 @@ const updateReviews: RequestHandler = async (req, res, next) => {
 
 const likeReviews: RequestHandler = async (req, res, next) => {
     try {
-        const reviewOwner = await Review.findByPk(req.params.reviewId, {
-            attributes: ['user_id'],
-        });
-        if (reviewOwner?.dataValues.user_id == user_id) {
+        if (await isUserOwnerOfReview(user_id, req.params.reviewId)) {
             const error = new Error(
                 '자신의 리뷰에는 좋아요를 누를 수 없습니다.'
             );
